Complete mirror from docker/ to root directories and remove redundant docker/ folder (#2)
* Checkpoint from VS Code for coding agent session

* Initial plan

* Complete mirror by copying missing content and fixing paths


* Successfully remove docker/ folder - mirroring complete


* Merge commit '95eadf32968f23612e563d17f214333a8c46ac50' into copilot/vscode1757077632051

* chore: update script paths in package-lock.json

- Changed paths for fvtt-compose-gen and fvtt-pod scripts
- Updated from 'docker/scripts/' to 'scripts/' for better organization

* Refactor package.json: Move jest to devDependencies and add Babel dependencies
