+++ patches/common/XX-patch-entrypoint.sh.template
