+++ tests/unit/patches/common/helpers/componentInstaller.unit.test.js
@@ -4,7 +4,7 @@ import http from 'node:http';
 import zlib from 'node:zlib';
 
 // eslint-disable-next-line import/no-commonjs
-const { ComponentInstaller } = require('../../../../../docker/patches/common/helpers/componentInstaller.mjs');
+const { ComponentInstaller } = require('../../../../../patches/common/helpers/componentInstaller.mjs');
 
 function padToBlockSize(len, block = 512) { return (block - (len % block)) % block; }
 function buildTar(entries) {
@@ -37,7 +37,7 @@ function buildTar(entries) {
   return Buffer.concat(chunks);
 }
 
-describe('docker/patches/common/helpers/componentInstaller.mjs', () => {
+describe('patches/common/helpers/componentInstaller.mjs', () => {
   const tmpRoot = path.join(process.cwd(), '.jest-tmp-installer');
   const dataDir = path.join(tmpRoot, 'Data');
   const cacheDir = path.join(tmpRoot, 'cache');
@@ -163,7 +163,7 @@ describe('docker/patches/common/helpers/componentInstaller.mjs', () => {
       PATCH_FORCE_NODE_EXTRACT: '1'
     };
   // eslint-disable-next-line import/no-commonjs
-  const { ComponentInstaller } = require('../../../../../docker/patches/common/helpers/componentInstaller.mjs');
+  const { ComponentInstaller } = require('../../../../../patches/common/helpers/componentInstaller.mjs');
     const spy = jest.spyOn(process, 'exit').mockImplementation((code) => { throw new Error(`exit:${code}`); });
     try {
       // Constructor will attempt to read config and exit
@@ -183,7 +183,7 @@ describe('docker/patches/common/helpers/componentInstaller.mjs', () => {
     fs.writeFileSync(cfgPath, JSON.stringify(cfg));
     const env = { FOUNDRY_DATA_DIR: dataDir, CONTAINER_CONFIG_PATH: cfgPath, COMPONENT_CACHE: cacheDir };
   // eslint-disable-next-line import/no-commonjs
-  const { ComponentInstaller } = require('../../../../../docker/patches/common/helpers/componentInstaller.mjs');
+  const { ComponentInstaller } = require('../../../../../patches/common/helpers/componentInstaller.mjs');
     const spy = jest.spyOn(process, 'exit').mockImplementation((code) => { throw new Error(`exit:${code}`); });
     try {
       // @ts-ignore
@@ -202,7 +202,7 @@ describe('docker/patches/common/helpers/componentInstaller.mjs', () => {
     fs.writeFileSync(cfgPath, JSON.stringify(cfg));
     const env = { FOUNDRY_VERSION: '13.100', FOUNDRY_DATA_DIR: dataDir, CONTAINER_CONFIG_PATH: cfgPath, COMPONENT_CACHE: cacheDir };
   // eslint-disable-next-line import/no-commonjs
-  const { ComponentInstaller } = require('../../../../../docker/patches/common/helpers/componentInstaller.mjs');
+  const { ComponentInstaller } = require('../../../../../patches/common/helpers/componentInstaller.mjs');
     const spy = jest.spyOn(process, 'exit').mockImplementation((code) => { throw new Error(`exit:${code}`); });
     try {
       // @ts-ignore
@@ -221,7 +221,7 @@ describe('docker/patches/common/helpers/componentInstaller.mjs', () => {
     fs.writeFileSync(cfgPath, JSON.stringify(cfg));
     const env = { FOUNDRY_VERSION: '13.0.0', FOUNDRY_DATA_DIR: dataDir, CONTAINER_CONFIG_PATH: cfgPath, COMPONENT_CACHE: cacheDir };
   // eslint-disable-next-line import/no-commonjs
-  const { ComponentInstaller } = require('../../../../../docker/patches/common/helpers/componentInstaller.mjs');
+  const { ComponentInstaller } = require('../../../../../patches/common/helpers/componentInstaller.mjs');
     const spy = jest.spyOn(process, 'exit').mockImplementation((code) => { throw new Error(`exit:${code}`); });
     try {
       // @ts-ignore
